+++ src/PdfPoint.php
@@ -27,16 +27,6 @@ class PdfPoint
     }
 
     /**
-     * Returns these x and y values as an array.
-     *
-     * @return array{0: float, 1: float}
-     */
-    public function asArray(): array
-    {
-        return [$this->x, $this->y];
-    }
-
-    /**
      * Returns if the given point is equal to this instance.
      *
      * To be equal, the abscissa and the ordinate values must be equal